# api/01-openapi-spec.yaml
openapi: 3.0.3
info:
  title: Ride Circle API
  description: |
    Comprehensive API for the Ride Circle peer-to-peer car rental platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute
    - Admin endpoints: 5000 requests/minute
    
    ## Multi-tenancy
    All requests must include the tenant context via subdomain (mumbai.ridecircle.in) 
    or X-Tenant-Code header.
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional context.
    
  version: 1.0.0
  contact:
    name: Ride Circle API Support
    email: api-support@ridecircle.in
  license:
    name: Proprietary
servers:
  - url: https://api.ridecircle.in/v1
    description: Production server
  - url: https://staging-api.ridecircle.in/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /users/kyc:
    post:
      tags: [User Management]
      summary: Upload KYC document
      description: Upload KYC document for verification
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [documentType, documentNumber, file]
              properties:
                documentType:
                  type: string
                  enum: [aadhaar, pan, driving_license, passport]
                documentNumber:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCDocument'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [User Management]
      summary: Get KYC status
      description: Get current user's KYC verification status
      responses:
        '200':
          description: KYC status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/KYCDocument'
                  overallStatus:
                    type: string
                    enum: [pending, in_review, approved, rejected]
                  completionPercentage:
                    type: number
                    minimum: 0
                    maximum: 100

  # =====================================================
  # VEHICLE MANAGEMENT
  # =====================================================
...
  /bookings:
    get:
      tags: [Bookings]
      summary: Get user bookings
      description: Get bookings for current user (as customer or host)
      parameters:
        - name: role
          in: query
          description: Filter by user role in booking
          schema:
            type: string
            enum: [customer, host]
        - name: status
          in: query
          description: Filter by booking status
          schema:
            type: string
            enum: [pending, confirmed, payment_pending, active, completed, cancelled, no_show, disputed]
        - name: startDate
          in: query
          description: Filter bookings from this date
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter bookings until this date
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'

    post:
      tags: [Bookings]
      summary: Create booking
      description: Create a new booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Vehicle not available for selected dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /payments/process:
    post:
      tags: [Payments]
      summary: Process payment
      description: Process payment for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, paymentMethod, amount]
              properties:
                bookingId:
                  type: string
                  format: uuid
                paymentMethod:
                  type: string
                  enum: [card, upi, netbanking, wallet]
                amount:
                  type: number
                  minimum: 0
                transactionType:
                  type: string
                  enum: [booking, security_deposit]
                  default: booking
                gatewayData:
                  type: object
                  description: Gateway-specific payment data
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Payment failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        errors:
          type: array
          items:
            type: string
        code:
          type: string
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false

    # Authentication schemas
...
    KYCDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        documentType:
          type: string
          enum: [aadhaar, pan, driving_license, passport]
        documentNumber:
          type: string
        documentUrl:
          type: string
          format: uri
        verificationStatus:
          type: string
          enum: [pending, in_review, approved, rejected, expired]
        verifiedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # Vehicle schemas
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingNumber:
          type: string
          example: "RC2024001234"
        customerId:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, payment_pending, active, completed, cancelled, no_show, disputed]
        baseAmount:
          type: number
        taxes:
          type: number
        fees:
          type: number
        discounts:
          type: number
        totalAmount:
          type: number
        securityDeposit:
          type: number
        paymentStatus:
          type: string
        pickupLocation:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        dropoffLocation:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        actualStartTime:
          type: string
          format: date-time
        actualEndTime:
          type: string
          format: date-time
        startOdometer:
          type: integer
        endOdometer:
          type: integer
        fuelLevelStart:
          type: integer
        fuelLevelEnd:
          type: integer
        specialRequests:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingCreate:
      type: object
      required: [vehicleId, startDate, endDate]
      properties:
        vehicleId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        pickupLocation:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        dropoffLocation:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        specialRequests:
          type: string
          maxLength: 500
...
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # Payment schemas
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
          example: "INR"
        paymentMethod:
          type: string
          enum: [card, upi, netbanking, wallet, cash]
        transactionType:
          type: string
          enum: [booking, security_deposit, refund, penalty, commission]
        status:
          type: string
          enum: [pending, processing, completed, failed, refunded, disputed]
        gatewayProvider:
          type: string
        gatewayTransactionId:
          type: string
        processedAt:
          type: string
          format: date-time
        failedReason:
          type: string
        refundAmount:
          type: number
        refundedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Payment'
        gatewayData:
          type: object
          description: Gateway-specific response data
...
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Validation failed"
            errors: ["Email is required", "Phone number is invalid"]
            code: "VALIDATION_ERROR"
            timestamp: "2024-01-15T10:30:00Z"
...