# app/api/v1/bookings/route.ts
import { type NextRequest, NextResponse } from "next/server"
import { z } from "zod"
import { nanoid } from "nanoid"

const BookingCreateSchema = z.object({
  vehicleId: z.string().uuid(),
  startDate: z.string().datetime(),
  endDate: z.string().datetime(),
  pickupLocation: z
    .object({
      lat: z.number(),
      lng: z.number(),
    })
    .optional(),
  dropoffLocation: z
    .object({
      lat: z.number(),
      lng: z.number(),
    })
    .optional(),
  specialRequests: z.string().max(500).optional(),
})

// GET /api/v1/bookings - Get user bookings
export async function GET(request: NextRequest) {
  try {
    // TODO: Verify JWT token and extract user ID
    const authHeader = request.headers.get("authorization")
    if (!authHeader) {
      return NextResponse.json(
        {
          success: false,
          message: "Authentication required",
        },
        { status: 401 },
      )
    }

    const { searchParams } = new URL(request.url)
    const status = searchParams.get("status")
    const page = Number(searchParams.get("page")) || 1
    const limit = Number(searchParams.get("limit")) || 10

    // TODO: Get bookings from database for current user
    // Mock booking data for demonstration
    const mockBookings = [
      {
        id: "booking-1",
        bookingNumber: "RC2024001234",
        customerId: "current-user-id",
        hostId: "host-1",
        vehicleId: "vehicle-1",
        vehicle: {
          make: "Honda",
          model: "City",
          year: 2020,
          registrationNumber: "MH01AB1234",
          images: ["/honda-city-white-car.png"],
        },
        host: {
          name: "Rajesh Kumar",
          phone: "+919876543210",
          rating: 4.8,
        },
        startDate: "2024-01-20T09:00:00Z",
        endDate: "2024-01-22T18:00:00Z",
        status: "confirmed",
        totalAmount: 5000,
        securityDeposit: 5000,
        paymentStatus: "paid",
        createdAt: "2024-01-15T10:30:00Z",
        pickupLocation: {
          lat: 19.076,
          lng: 72.8777,
          address: "123 Main Street, Bandra West, Mumbai",
        },
        dropoffLocation: {
          lat: 19.076,
          lng: 72.8777,
          address: "123 Main Street, Bandra West, Mumbai",
        },
      },
    ]

    console.log("[v0] Bookings retrieved:", { userId: "current-user-id", status, page, limit })

    return NextResponse.json({
      success: true,
      message: "Bookings retrieved successfully",
      data: {
        bookings: mockBookings,
        pagination: {
          page,
          limit,
          total: mockBookings.length,
          totalPages: 1,
        },
      },
    })
  } catch (error) {
    console.error("[v0] Get bookings error:", error)
    return NextResponse.json(
      {
        success: false,
        message: "Failed to retrieve bookings",
      },
      { status: 500 },
    )
  }
}

// POST /api/v1/bookings - Create booking
export async function POST(request: NextRequest) {
  try {
    // TODO: Verify JWT token and extract user ID
    const authHeader = request.headers.get("authorization")
    if (!authHeader) {
      return NextResponse.json(
        {
          success: false,
          message: "Authentication required",
        },
        { status: 401 },
      )
    }

    const body = await request.json()
    const validatedData = BookingCreateSchema.parse(body)

    // TODO: Check vehicle availability
    // TODO: Calculate pricing
    // TODO: Validate dates

    const startDate = new Date(validatedData.startDate)
    const endDate = new Date(validatedData.endDate)
    const durationHours = Math.ceil((endDate.getTime() - startDate.getTime()) / (1000 * 60 * 60))
    const durationDays = Math.ceil(durationHours / 24)

    // Mock pricing calculation
    const basePricePerDay = 2500
    const baseAmount = basePricePerDay * durationDays
    const platformFee = Math.round(baseAmount * 0.05) // 5% platform fee
    const gst = Math.round((baseAmount + platformFee) * 0.18) // 18% GST
    const totalAmount = baseAmount + platformFee + gst
    const securityDeposit = 5000

    // Create booking record
    const booking = {
      id: nanoid(),
      bookingNumber: `RC${new Date().getFullYear()}${Math.floor(Math.random() * 1000000)
        .toString()
        .padStart(6, "0")}`,
      customerId: "current-user-id", // Would come from JWT token
      hostId: "host-1", // Would come from vehicle data
      vehicleId: validatedData.vehicleId,
      startDate: validatedData.startDate,
      endDate: validatedData.endDate,
      pickupLocation: validatedData.pickupLocation,
      dropoffLocation: validatedData.dropoffLocation,
      specialRequests: validatedData.specialRequests,
      status: "pending",
      pricing: {
        baseAmount,
        platformFee,
        gst,
        totalAmount,
        securityDeposit,
        durationDays,
        durationHours,
      },
      paymentStatus: "pending",
      createdAt: new Date(),
      updatedAt: new Date(),
    }

    // TODO: Save booking to database
    // TODO: Create payment intent
    // TODO: Send notification to host
    console.log("[v0] Booking created:", { bookingId: booking.id, customerId: booking.customerId })

    return NextResponse.json(
      {
        success: true,
        message: "Booking created successfully",
        data: {
          booking: {
            id: booking.id,
            bookingNumber: booking.bookingNumber,
            status: booking.status,
            totalAmount: booking.pricing.totalAmount,
            securityDeposit: booking.pricing.securityDeposit,
            paymentStatus: booking.paymentStatus,
            createdAt: booking.createdAt,
          },
          nextSteps: {
            paymentRequired: true,
            hostApprovalRequired: true,
            estimatedConfirmationTime: "2-4 hours",
          },
        },
      },
      { status: 201 },
    )
  } catch (error) {
    if (error instanceof z.ZodError) {
      return NextResponse.json(
        {
          success: false,
          message: "Validation failed",
          errors: error.errors,
        },
        { status: 400 },
      )
    }

    console.error("[v0] Booking creation error:", error)
    return NextResponse.json(
      {
        success: false,
        message: "Booking creation failed",
      },
      { status: 500 },
    )
  }
}