# app/api/v1/payments/process/route.ts
import { type NextRequest, NextResponse } from "next/server"
import { z } from "zod"
import { nanoid } from "nanoid"

const PaymentProcessSchema = z.object({
  bookingId: z.string().uuid(),
  paymentMethod: z.enum(["card", "upi", "netbanking", "wallet"]),
  amount: z.number().min(1),
  currency: z.string().default("INR"),
  gatewayProvider: z.enum(["razorpay", "paytm"]).default("razorpay"),
  returnUrl: z.string().url().optional(),
  metadata: z.record(z.string()).optional(),
})

export async function POST(request: NextRequest) {
  try {
    // TODO: Verify JWT token and extract user ID
    const authHeader = request.headers.get("authorization")
    if (!authHeader) {
      return NextResponse.json(
        {
          success: false,
          message: "Authentication required",
        },
        { status: 401 },
      )
    }

    const body = await request.json()
    const validatedData = PaymentProcessSchema.parse(body)

    // TODO: Get booking details from database
    // TODO: Verify booking belongs to current user
    // TODO: Check if payment is already processed

    // Create payment record
    const payment = {
      id: nanoid(),
      bookingId: validatedData.bookingId,
      userId: "current-user-id", // Would come from JWT token
      amount: validatedData.amount,
      currency: validatedData.currency,
      paymentMethod: validatedData.paymentMethod,
      gatewayProvider: validatedData.gatewayProvider,
      status: "initiated",
      transactionType: "booking",
      createdAt: new Date(),
      metadata: validatedData.metadata,
    }

    // TODO: Create payment intent with gateway (Razorpay/Paytm)
    // Mock gateway response
    const gatewayResponse = {
      paymentId: `pay_${nanoid()}`,
      orderId: `order_${nanoid()}`,
      amount: validatedData.amount * 100, // Convert to paise for Razorpay
      currency: validatedData.currency,
      status: "created",
      paymentUrl: `https://checkout.razorpay.com/v1/checkout.js`,
      key: "rzp_test_key", // Would come from environment
    }

    // TODO: Save payment to database
    console.log("[v0] Payment initiated:", {
      paymentId: payment.id,
      bookingId: payment.bookingId,
      amount: payment.amount,
    })

    return NextResponse.json(
      {
        success: true,
        message: "Payment initiated successfully",
        data: {
          payment: {
            id: payment.id,
            status: payment.status,
            amount: payment.amount,
            currency: payment.currency,
            createdAt: payment.createdAt,
          },
          gateway: {
            provider: validatedData.gatewayProvider,
            paymentId: gatewayResponse.paymentId,
            orderId: gatewayResponse.orderId,
            paymentUrl: gatewayResponse.paymentUrl,
            key: gatewayResponse.key,
            amount: gatewayResponse.amount,
          },
          nextSteps: {
            redirectToGateway: true,
            webhookUrl: `${request.nextUrl.origin}/api/v1/webhooks/payments`,
          },
        },
      },
      { status: 201 },
    )
  } catch (error) {
    if (error instanceof z.ZodError) {
      return NextResponse.json(
        {
          success: false,
          message: "Validation failed",
          errors: error.errors,
        },
        { status: 400 },
      )
    }

    console.error("[v0] Payment processing error:", error)
    return NextResponse.json(
      {
        success: false,
        message: "Payment processing failed",
      },
      { status: 500 },
    )
  }
}