# app/api/v1/users/kyc/route.ts
import { type NextRequest, NextResponse } from "next/server"
import { z } from "zod"
import { nanoid } from "nanoid"

const KYCUploadSchema = z.object({
  documentType: z.enum(["aadhaar", "driving_license", "passport", "voter_id"]),
  documentNumber: z.string().min(1),
  frontImageUrl: z.string().url(),
  backImageUrl: z.string().url().optional(),
  selfieUrl: z.string().url(),
  digilockerConsent: z.boolean().default(false),
  videoDeclarationUrl: z.string().url().optional(),
})

// GET /api/v1/users/kyc - Get KYC status
export async function GET(request: NextRequest) {
  try {
    // TODO: Verify JWT token and extract user ID
    const authHeader = request.headers.get("authorization")
    if (!authHeader) {
      return NextResponse.json(
        {
          success: false,
          message: "Authentication required",
        },
        { status: 401 },
      )
    }

    // TODO: Get KYC status from database
    // Mock KYC data for demonstration
    const kycStatus = {
      userId: "current-user-id",
      status: "pending_review",
      submittedAt: "2024-01-15T10:30:00Z",
      reviewedAt: null,
      documents: [
        {
          type: "aadhaar",
          status: "verified",
          documentNumber: "XXXX-XXXX-1234",
          verifiedAt: "2024-01-15T11:00:00Z",
        },
        {
          type: "driving_license",
          status: "pending_review",
          documentNumber: "MH01-20240001234",
          submittedAt: "2024-01-15T10:30:00Z",
        },
      ],
      biometricVerification: {
        status: "completed",
        matchScore: 95.5,
        verifiedAt: "2024-01-15T10:45:00Z",
      },
      videoDeclaration: {
        status: "completed",
        recordedAt: "2024-01-15T10:35:00Z",
        duration: 45, // seconds
      },
      nextSteps: ["Driving license verification in progress", "Final admin review pending"],
    }

    console.log("[v0] KYC status retrieved:", { userId: "current-user-id" })

    return NextResponse.json({
      success: true,
      message: "KYC status retrieved successfully",
      data: kycStatus,
    })
  } catch (error) {
    console.error("[v0] Get KYC status error:", error)
    return NextResponse.json(
      {
        success: false,
        message: "Failed to retrieve KYC status",
      },
      { status: 500 },
    )
  }
}

// POST /api/v1/users/kyc - Upload KYC documents
export async function POST(request: NextRequest) {
  try {
    // TODO: Verify JWT token and extract user ID
    const authHeader = request.headers.get("authorization")
    if (!authHeader) {
      return NextResponse.json(
        {
          success: false,
          message: "Authentication required",
        },
        { status: 401 },
      )
    }

    const body = await request.json()
    const validatedData = KYCUploadSchema.parse(body)

    // Create KYC record
    const kycRecord = {
      id: nanoid(),
      userId: "current-user-id", // Would come from JWT token
      documentType: validatedData.documentType,
      documentNumber: validatedData.documentNumber,
      frontImageUrl: validatedData.frontImageUrl,
      backImageUrl: validatedData.backImageUrl,
      selfieUrl: validatedData.selfieUrl,
      digilockerConsent: validatedData.digilockerConsent,
      videoDeclarationUrl: validatedData.videoDeclarationUrl,
      status: "submitted",
      submittedAt: new Date(),
      createdAt: new Date(),
    }

    // TODO: Save KYC record to database
    // TODO: Trigger document verification process
    // TODO: If DigiLocker consent, initiate DigiLocker verification
    // TODO: Trigger biometric face matching
    // TODO: Send notification to admin for review

    console.log("[v0] KYC documents uploaded:", {
      kycId: kycRecord.id,
      userId: kycRecord.userId,
      documentType: kycRecord.documentType,
    })

    // Mock verification process initiation
    const verificationTasks = []

    if (validatedData.digilockerConsent) {
      verificationTasks.push("DigiLocker verification initiated")
    }

    verificationTasks.push("Document OCR processing started")
    verificationTasks.push("Biometric face matching queued")

    if (validatedData.videoDeclarationUrl) {
      verificationTasks.push("Video declaration analysis started")
    }

    return NextResponse.json(
      {
        success: true,
        message: "KYC documents uploaded successfully",
        data: {
          kycRecord: {
            id: kycRecord.id,
            status: kycRecord.status,
            documentType: kycRecord.documentType,
            submittedAt: kycRecord.submittedAt,
          },
          verification: {
            estimatedProcessingTime: "2-4 hours",
            tasksInitiated: verificationTasks,
            trackingId: kycRecord.id,
          },
          nextSteps: [
            "Document verification in progress",
            "You will be notified once review is complete",
            "Estimated completion: 2-4 hours",
          ],
        },
      },
      { status: 201 },
    )
  } catch (error) {
    if (error instanceof z.ZodError) {
      return NextResponse.json(
        {
          success: false,
          message: "Validation failed",
          errors: error.errors,
        },
        { status: 400 },
      )
    }

    console.error("[v0] KYC upload error:", error)
    return NextResponse.json(
      {
        success: false,
        message: "KYC upload failed",
      },
      { status: 500 },
    )
  }
}